+++ src/commands/image/updateImage/updateImage.ts
@@ -15,7 +15,7 @@ import { createActivityContext } from "../../../utils/activity/activityUtils";
 import { localize } from "../../../utils/localize";
 import { pickContainerApp } from "../../../utils/pickItem/pickContainerApp";
 import { pickRevision, pickRevisionDraft } from "../../../utils/pickItem/pickRevision";
-import { getParentResourceFromItem, showRevisionDraftDeployPopup } from "../../../utils/revisionDraftUtils";
+import { getParentResourceFromItem } from "../../../utils/revisionDraftUtils";
 import type { ImageSourceBaseContext } from "../imageSource/ImageSourceBaseContext";
 import { ImageSourceListStep } from "../imageSource/ImageSourceListStep";
 import { UpdateImageDraftStep } from "./UpdateImageDraftStep";
@@ -74,6 +74,4 @@ export async function updateImage(context: IActionContext, node?: ContainerAppIt
 
     await wizard.prompt();
     await wizard.execute();
-
-    void showRevisionDraftDeployPopup(context, containerApp);
 }
