+++ src/commands/scaling/scaleRule/addScaleRule/addScaleRule.ts
@@ -10,7 +10,7 @@ import type { ScaleRuleGroupItem } from "../../../../tree/scaling/ScaleRuleGroup
 import { createActivityContext } from "../../../../utils/activity/activityUtils";
 import { localize } from "../../../../utils/localize";
 import { pickScaleRuleGroup } from "../../../../utils/pickItem/pickScale";
-import { getParentResource, showRevisionDraftDeployPopup } from "../../../../utils/revisionDraftUtils";
+import { getParentResource } from "../../../../utils/revisionDraftUtils";
 import { AddScaleRuleStep } from "./AddScaleRuleStep";
 import type { IAddScaleRuleContext } from "./IAddScaleRuleContext";
 import { ScaleRuleNameStep } from "./ScaleRuleNameStep";
@@ -41,6 +41,4 @@ export async function addScaleRule(context: IActionContext, node?: ScaleRuleGrou
     await wizard.prompt();
     wizardContext.activityTitle = localize('addScaleRuleTitle', 'Add {0} rule "{1}" to "{2}" (draft)', wizardContext.newRuleType, wizardContext.newRuleName, parentResource.name);
     await wizard.execute();
-
-    void showRevisionDraftDeployPopup(context, containerApp);
 }
