+++ src/commands/scaling/scaleRule/deleteScaleRule/DeleteScaleRuleStep.ts
@@ -17,20 +17,20 @@ export class DeleteScaleRuleStep<T extends ScaleRuleContext> extends RevisionDra
         super(baseItem);
     }
 
-    public async execute(context: ScaleRuleContext): Promise<void> {
+    public async execute(context: T): Promise<void> {
         this.revisionDraftTemplate.scale ||= {};
         this.revisionDraftTemplate.scale.rules ||= [];
 
         const index = this.revisionDraftTemplate.scale.rules.findIndex(r => r.name === nonNullValueAndProp(context.scaleRule, 'name'));
         this.revisionDraftTemplate.scale.rules.splice(index, 1);
 
-        this.updateRevisionDraftWithTemplate();
+        await this.updateRevisionDraftWithTemplate(context);
 
         const resourceName = getParentResourceFromItem(this.baseItem).name;
         ext.outputChannel.appendLog(localize('deletedScaleRule', 'Deleted rule "{0}" to "{1}" (draft)', context.scaleRule?.name, resourceName));
     }
 
-    public shouldExecute(context: ScaleRuleContext): boolean {
+    public shouldExecute(context: T): boolean {
         return !!context.scaleRule;
     }
 }
