+++ src/commands/scaling/scaleRule/deleteScaleRule/deleteScaleRule.ts
@@ -10,7 +10,7 @@ import type { ScaleRuleItem } from "../../../../tree/scaling/ScaleRuleItem";
 import { createActivityContext } from "../../../../utils/activity/activityUtils";
 import { localize } from "../../../../utils/localize";
 import { pickScaleRuleGroup } from "../../../../utils/pickItem/pickScale";
-import { getParentResource, showRevisionDraftDeployPopup } from "../../../../utils/revisionDraftUtils";
+import { getParentResource } from "../../../../utils/revisionDraftUtils";
 import type { ScaleRuleContext } from "../ScaleRuleContext";
 import { ScaleRuleListStep } from "../ScaleRuleListStep";
 import { DeleteScaleRuleStep } from "./DeleteScaleRuleStep";
@@ -49,6 +49,4 @@ export async function deleteScaleRule(context: IActionContext, node?: ScaleRuleI
     await wizard.prompt();
     wizardContext.activityTitle = localize('deleteScaleRuleTitle', 'Delete rule "{0}" from "{1}" (draft)', wizardContext.scaleRule?.name, parentResource.name);
     await wizard.execute();
-
-    void showRevisionDraftDeployPopup(context, containerApp);
 }
